% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{cFolder}
\alias{cFolder}
\alias{cFolder,roCache_versionBase-method}
\title{cFolder}
\usage{
cFolder(cacheHandle, ...)

\S4method{cFolder}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
cFolder
}
