% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultFuncHandle.R
\name{defaultFuncHandle}
\alias{defaultFuncHandle}
\title{defaultFuncHandle}
\usage{
defaultFuncHandle(getDigests = FALSE, object)
}
\arguments{
\item{getDigests}{(TRUE/FALSE), return the list or input digest, or return object.}

\item{object}{The object to return (and store in cache).}
}
\description{
The default function handle used for \code{\link{storeCache}}.
The func handle has one argument names object, reflecting the
object to store.
}
\examples{
\dontrun{
storeCache(cacheHdl, id = "ID1", object="ABCD")
}
}
