% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{dFile}
\alias{dFile}
\alias{dFile,roCache_versionBase-method}
\title{dFile}
\usage{
dFile(cacheHandle, ...)

\S4method{dFile}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
dFile
}
