% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digestOfFunc.R
\name{digestOfFunc}
\alias{digestOfFunc}
\title{Get a digest of a function body}
\usage{
digestOfFunc(f)
}
\arguments{
\item{f}{the function to obtain the digest of}
}
\value{
A character string representing the functions digest
}
\description{
The digest of the function is excluding comment or whitespace structure.
}
\examples{

aFunc1<-function(){
  return("aValue")
}

aFunc2<-function(){
#Additional whitespce 
  return("aValue" ) ##This is a comment
}

digestOfFunc(aFunc1)
digestOfFunc(aFunc2)

}
