% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{get_cFolder}
\alias{get_cFolder}
\alias{get_cFolder,roCache_versionBase-method}
\alias{get_cFolder.roCache_versionBase}
\title{getter: get_cFolder}
\usage{
get_cFolder(cacheHandle, ...)

\S4method{get_cFolder}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
A getter to obtain the cFolder from the cacheHandle.
}
