% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{get_dFile}
\alias{get_dFile}
\alias{get_dFile,roCache_versionBase-method}
\alias{get_dFile.roCache_versionBase}
\title{getter: get_dFile}
\usage{
get_dFile(cacheHandle, ...)

\S4method{get_dFile}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
A getter to obtain the dFile from the cacheHandle.
}
