% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{get_prevVersion}
\alias{get_prevVersion}
\alias{get_prevVersion,roCache_versionBase-method}
\alias{get_prevVersion.roCache_versionBase}
\title{getter: get_prevVersion}
\usage{
get_prevVersion(cacheHandle, ...)

\S4method{get_prevVersion}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
A getter to obtain the prevVersion from the cacheHandle.
}
