% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{get_syncAllowed}
\alias{get_syncAllowed}
\alias{get_syncAllowed,roCache_versionBase-method}
\alias{get_syncAllowed.roCache_versionBase}
\title{getter: get_syncAllowed}
\usage{
get_syncAllowed(cacheHandle, ...)

\S4method{get_syncAllowed}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
A getter to obtain 'syncAllowed' status from the cacheHandle.
}
