% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{get_version}
\alias{get_version}
\alias{get_version,roCache_versionBase-method}
\alias{get_version.roCache_versionBase}
\title{getter: get_version}
\usage{
get_version(cacheHandle, ...)

\S4method{get_version}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
A getter to obtain the version from the cacheHandle.
}
