% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{overwrite}
\alias{overwrite}
\alias{overwrite<-}
\alias{overwrite,roCache_versionBase-method}
\alias{overwrite<-,roCache_versionBase,logical-method}
\title{overwrite}
\usage{
overwrite(cacheHandle, ...)

overwrite(cacheHandle) <- value

\S4method{overwrite}{roCache_versionBase}(cacheHandle, ...)

\S4method{overwrite}{roCache_versionBase,logical}(cacheHandle) <- value
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}

\item{value}{The value to set}
}
\description{
overwrite
}
