% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{prevVersion}
\alias{prevVersion}
\alias{prevVersion,roCache_versionBase-method}
\title{prevVersion}
\usage{
prevVersion(cacheHandle, ...)

\S4method{prevVersion}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
prevVersion
}
