% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{rdsVersion}
\alias{rdsVersion}
\alias{rdsVersion,roCache_versionBase-method}
\title{rdsVersion}
\usage{
rdsVersion(cacheHandle, ...)

\S4method{rdsVersion}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
rdsVersion
}
