% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_001_GenericFunctions.R,
%   R/aaa_070_class_roCache_versionBase.R
\name{version}
\alias{version}
\alias{version,roCache_versionBase-method}
\title{version}
\usage{
version(cacheHandle, ...)

\S4method{version}{roCache_versionBase}(cacheHandle, ...)
}
\arguments{
\item{cacheHandle}{The handle of interest.}

\item{...}{not used}
}
\description{
version
}
